requirements.txt
python-telegram-bot
requests
websockets
